More separation of tests
