+++ spezi-invitation-code/tests/invitationCodeVerifier.test.js
@@ -11,21 +11,6 @@ const firebaseTest = require("firebase-functions-test")();
 const InvitationCodeVerifier = require("../index.js");
 const {https} = require("firebase-functions/v2");
 
-jest.mock("firebase-admin", () => {
-  const firestore = {
-    doc: jest.fn(),
-    collection: jest.fn(),
-    runTransaction: jest.fn(),
-  };
-  return {
-    initializeApp: jest.fn(),
-    firestore: jest.fn(() => firestore),
-    app: jest.fn(() => ({
-      delete: jest.fn(),
-    })),
-  };
-});
-
 describe("InvitationCodeVerifier", () => {
   let verifier;
   let firestore;
@@ -57,29 +42,10 @@ describe("InvitationCodeVerifier", () => {
   });
 
   test("should validate user invitation code successfully", async () => {
-    firestore.collection.mockReturnValue({
-      where: jest.fn().mockReturnThis(),
-      limit: jest.fn().mockReturnThis(),
-      get: jest.fn().mockResolvedValue({
-        empty: false,
-        docs: [{id: "validCode"}],
-      }),
-    });
-    firestore.doc.mockReturnValueOnce({
-      get: jest.fn().mockResolvedValue({
-        exists: true,
-        data: () => ({invitationCode: "validCode"}),
-      }),
-    });
-
     await expect(verifier.validateUserInvitationCode("HNzc8VN8maeT1uUnABgWozWMPT6x")).resolves.toBeUndefined();
   });
 
   test("should throw an error if invitationCode does not exist or already used", async () => {
-    firestore.doc.mockReturnValueOnce({
-      get: jest.fn().mockResolvedValue({exists: false}),
-    });
-
     await expect(verifier.enrollUserInStudy("HNzc8VN8maeT1uUnABgWozWMPT6x", "3Op7vweq")).rejects.toThrow(
         new https.HttpsError(
             "not-found",
@@ -90,30 +56,12 @@ describe("InvitationCodeVerifier", () => {
 
   describe("validateUserInvitationCode", () => {
     test("should throw an error if no valid invitation code found for the user", async () => {
-      firestore.collection.mockReturnValue({
-        where: jest.fn().mockReturnThis(),
-        limit: jest.fn().mockReturnThis(),
-        get: jest.fn().mockResolvedValue({empty: true}),
-      });
-
       await expect(verifier.validateUserInvitationCode("user123")).rejects.toThrow(
           new https.HttpsError("not-found", "No valid invitation code found for user user123."),
       );
     });
 
     test("should throw an error if user document does not exist or contains incorrect invitation code", async () => {
-      firestore.collection.mockReturnValue({
-        where: jest.fn().mockReturnThis(),
-        limit: jest.fn().mockReturnThis(),
-        get: jest.fn().mockResolvedValue({
-          empty: false,
-          docs: [{id: "validCode"}],
-        }),
-      });
-      firestore.doc.mockReturnValueOnce({
-        get: jest.fn().mockResolvedValue({exists: false}),
-      });
-
       await expect(verifier.validateUserInvitationCode("user123")).rejects.toThrow(
           new https.HttpsError("failed-precondition", "User document does not exist or contains incorrect invitation code."),
       );
@@ -133,19 +81,19 @@ describe("InvitationCodeVerifier", () => {
         ),
     );
   });
-});
-/*
-test("not sure what this tests actually", () => {
-  firestore.doc.mockReturnValueOnce({
-    get: jest.fn().mockResolvedValue({exists: true, data: () => ({used: false})}),
-  });
-  firestore.doc.mockReturnValueOnce({
-    get: jest.fn().mockResolvedValue({exists: false}),
-  });
-  firestore.runTransaction.mockImplementationOnce(async (updateFunction) => {
-    await updateFunction({
-      set: jest.fn(),
-      update: jest.fn(),
+
+  test("should not overwrite existing info?", () => {
+    firestore.doc.mockReturnValueOnce({
+      get: jest.fn().mockResolvedValue({exists: true, data: () => ({used: false})}),
+    });
+    firestore.doc.mockReturnValueOnce({
+      get: jest.fn().mockResolvedValue({exists: false}),
+    });
+    firestore.runTransaction.mockImplementationOnce(async (updateFunction) => {
+      await updateFunction({
+        set: jest.fn(),
+        update: jest.fn(),
+      });
     });
   });
-});*/
+});
